pyproject.toml
[project]
name = "incident-commander"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai-slim[openrouter]>=1.62.0",
]
